added EDIT SCREEN

+++ app/src/main/java/com/example/culinarapp/domain/usecases/AddRecipeUseCase.kt
@@ -7,6 +7,6 @@ import javax.inject.Inject
 class AddRecipeUseCase @Inject constructor(private val recipeRepository: RecipeRepository) {
     suspend operator fun invoke(recipe: Recipe): Recipe {
         val generatedId = recipeRepository.addRecipe(recipe)
-        return recipe.copy(id = generatedId.toInt())
+        return recipe.copy(id = generatedId)
     }
 }
